requirement.txt
streamlit==1.28.0
bcrypt==4.0.1
python-docx==0.8.11
PyPDF2==3.0.1
scikit-learn==1.3.2
mysql-connector-python==8.0.33